Match expression (#1)
* add match expression documentation

* Update docusaurus version

I need to use the admonitions feature of docusaurus
that is available only in the newer versions.
In this version, the font is little bigger.
Ther was two problems, the links param in navbar config
has been changed to items, and php is not loaded by default
for prism. I fixed those problems.